+++ src/components/Authentication/layouts/AuthLayout.tsx
@@ -4,14 +4,19 @@ import { type ReactNode, type PropsWithChildren } from "react";
 
 interface Props extends PropsWithChildren {
   Title: string | ReactNode;
+  imagePath?: string;
 }
-export default function AuthLayout({children,Title }: Props) {
+export default function AuthLayout({
+  imagePath = "/assets/signup_page.png",
+  children,
+  Title,
+}: Props) {
   return (
     <div>
-      <div className="flex w-full items-stretch justify-center">
+      <div className="flex w-full min-h-screen items-stretch justify-center">
         <div className="relative hidden flex-1 md:block">
           <StaticImage
-            src={"/assets/signup_page.png"}
+            src={imagePath}
             alt="Logo"
             fill
             className="object-cover"
